fix(app): invalid generateMetadata export build error
